third_party/WebKit/LayoutTests/fast/events/pointerevents/pointer-use-count.html

<!DOCTYPE HTML>
<script src="../../../resources/js-test.js"></script>
<style>
div.box {
  margin: 5px;
  padding: 20px;
  float: left;
  width: 50px;
  height: 50px;
}
</style>

<div id="grey" class="box" style="background-color:grey">
</div>

<div id="console"></div>

<script>
description("Verifies that use of pointerEvents is measured by use counters.");


var rect = document.getElementById("grey").getBoundingClientRect();
var x1 = rect.left + 5;
var y1 = rect.top + 5;

var PointerEventAttributeCount = 1321; // Comes from enum Feature in UseCounter.h
var PointerEventAddListenerCount = 1347; // Comes from enum Feature in UseCounter.h

function init() {
  var targetDiv = document.getElementById("grey");
  targetDiv.addEventListener("pointerdown", function(event) {
    debug("grey" + " received " + event.type);
    debug("    **** read pointerID to trigger MeasureAs *****");
    debug("pointerID is "+ event.pointerId);
  });
}

function runTest() {
  debug("    **** Move to grey box & mouse press *****");
  eventSender.mouseMoveTo(x1, y1);
  eventSender.mouseDown(1);
  
  shouldBe('internals.isUseCounted(document, PointerEventAttributeCount)', 'true');
  shouldBe('internals.isUseCounted(document, PointerEventAddListenerCount)', 'true');
}

init();
if (window.eventSender)
  runTest();
else
  debug("This test requires eventSender");

</script>
