third_party/WebKit/LayoutTests/fast/events/pointerevents/pointer-use-count-expected.txt

Verifies that use of pointerEvents is measured by use counters.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


    **** Move to grey box & mouse press *****
grey received pointerdown
    **** read pointerID to trigger MeasureAs *****
pointerID is 1
PASS internals.isUseCounted(document, PointerEventAttributeCount) is true
PASS internals.isUseCounted(document, PointerEventAddListenerCount) is true
PASS successfullyParsed is true

TEST COMPLETE
